REFACTOR: modularize langgraph
